[SYCLomatic] Include cmath when meet math API from standard library
